requirements.txt
absl-py==2.0.0
aniso8601==9.0.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
flasgger==0.9.7.1
Flask==2.3.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
Flask-MySQLdb==2.0.0
flask-restplus==0.13.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
flatbuffers==23.5.26
gast==0.5.4
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jsonschema-specifications==2023.11.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mysqlclient==2.2.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.7.0
PyMySQL==1.1.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
urllib3==2.1.0
Werkzeug==2.3.2
wrapt==1.14.1